refactor(desktop): move message-search types to searchTypes.ts
types.ts sat at its 1058-line size-gate ceiling, so the three new
search-operator fields tipped it to 1064. Relocating the message-search
types to a sibling module and re-exporting them keeps every import site
and the public type surface unchanged.
